Fix redacted read receipt responses
The current solution for edits and read receipts is not ideal. One
drawback is that redactions show up at the latest after logout/login and
reloading the room, even for those read receipt response events. The
workaround is to log a custom reason ('Read receipt response discarded
due to message edit.') and prevent displaying only of redactions that
have this reason set.