Fix: groups can be added later.
If a group is added to a room, and later another group should be added
(or even the old group removed), groups would be overwritten. This means
that users are kept in the room, but the state event only contains the
new group. This commit fixes this (in conjunction with a sdk change), so
that now single or multiple groups can be added later, not overwriting
the old groups.